requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
Django==5.2.1
django-ninja==1.4.3
factory_boy==3.3.3
Faker==37.4.0
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
PySocks==1.7.1
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
